Allow for intermittient failing of HLS playlist refresh download
At the moment when a playlist 404 happens the stream stalls and doesn't recover (you are just looking at a still image), attempting to seek will cause playback to end.

This fix stores a string copy of the playlist in the representation to pass in during prepareRepresentation in case of AdaptiveTree::download failing.

I've tested using a mitmproxy script to return 404 for the playlist (first 15 seconds of each minute) and had success.  Segment downloads and playback continue while the playlist is 404.