Make IA more resilient on missing subtitle fragments
We found that some streaming software offers a subtitle track, but may
not offer subtitle fragments (yet). This is usually done for live tv
where subtitles are not available for all broadcasted programmes, but
leaving out the subtitle track would require a full reload of the stream
to get subtitles when they are offered.

IA currently retries 10 times (11 tries) to get these unavailable
fragments causing an initial delay in playback for 15+ seconds.
And after the initial failure it does not retry future fragments.

Ideally an error on subtitle fragments should not be fatal to the
subtitle support and future segments should be retried. However IA would
need to redownload the manifest to support this.

This is documented in the original issue report which includes a study
of existing players and how they behave.

We found that the below behaviour works best:
- THEOplayer does not retry, but moves to the next subtitle fragment on
  error
- Shaka Player retries only once, and moves on to the next subtitle
  fragment

Other players failed out flat, or got stuck on retrying the same
subtitle fragment indefinitely.

So the next best thing is to drop subtitle support altogether on missing
subtitle fragments.