Update decrypter caps during InitializeDRM
Due to the HLS case we don't know the amount of DRM keys we will end up with, this can change with each new representation opened.
Update caps every time InitializeDRM is run. Thanks @matthuisman for
finding bug with streams that have no pssh and putting this in the
correct location.