Call ChooseNextRepresentation on periods > 1
Scenario:
Multi period stream starts, and due to previous bandwidth conditions the stream starts with a low bandwidth. Some small time passes and we're on the best rep available. When we transition to the next period we get given the low bandwidth again.
This change will make sure that before we get the first segment that we select an appropriate representation based on the download speeds so far.