Allow for playlist parsing to include crypto mode
Moved `CryptoMode` class out of AdaptiveDecrypter for other classes to easily include.
This is required as the CDM doesn't seem to like changing the mode after it has already performed decrypting in some situations (decrypt end in error). Now it can be set from the start of the stream rather than defaulting to CTR and later being changed with `ProcessMoof`.