Fix junk being sent as kid to decrypter
If we have a pssh without a default KID (DASHTree doesn't parse properly due to order of tags):
```
<ContentProtection schemeIdUri="urn:mpeg:dash:mp4protection:2011" value="cenc" cenc:default_KID="2384f91217d34162a995187704117759"/>
<ContentProtection schemeIdUri="urn:uuid:edef8ba9-79d6-4ace-a3c8-27dcd51d21ed"    >
  <cenc:pssh>AAAASXBzc2gAAAAA7e+LqXnWSs6jyCfc1R0h7QAAACkIARIQI4T5EhfTQWKplRh3BBF3WRoEaHVsdSIJMTcxMjQ5NzM3KgJIRA==</cenc:pssh>
</ContentProtection>
```
Then we end up sending a pointer to an empty string and crap gets sent to the decypter, then more bad stuff happens when `GetCapabilities` tries to use it.

Small clean-up included. With the full refactor of wvdecrypter we can abolish the `defkid` variable and just send the string instead.