[HLS] new periods pts reset to 0
I wrote a lengthy explanation in #676 about why this shoudn't happen - turns out that was incorrect and has since been solved in another way (correctly). New periods are often because of discontinuity, in any case it makes more sense to reset this to 0 especially since Kodi does not care for what the segment pts values are. Makes debugging easier and results in less confusion for others looking at the code and trying to debug.