Move update thread start to PostOpen
With current architecture in DASH we are always starting the update thread on the 'update tree' that is only used to extract information for our 'main' tree. Update tree only has a very short lifespan but the suspicion is that update tree is destroyed before the thread has started - update thread possibly then goes on to spawn its own update thread.